Module Sc_corpus

type 'r corpus

val make : workspace:Sc_core.Types.workspace -> 'raw_test Types.repr -> Types.params -> 'raw_test corpus Lwt.t

make ~workspace params creates a test corpus, represented as values of type raw_test.

val read_test : 'raw_test corpus -> _ Sc_sys.File.file -> 'raw_test option Lwt.t

read_test corpus f reads a test from file f; returns None if f does not contain a valid input.

val existing_test_ids : ?exclude:Basics.Digests.t -> _ corpus -> Basics.Digests.t Lwt.t

existing_test_ids ~exclude corpus is a stream of all identifiers of tests previously registered into corpus. Members of exclude, if given, are ignored.

val existing_tests : ?exclude:Basics.Digests.t -> ?sort_by_serial_num:bool -> 'raw_test corpus -> 'raw_test Types.test_view Lwt_stream.t

existing_tests ~exclude ~sort_by_serial_num corpus is a stream of all tests previously registered into corpus. Tests whose digest belongs to exclude, if given, are ignored. If sort_by_serial_num is true, sorts the tests in the stream by their serial number.

val share_test : ?on_share:(Stdlib.Digest.t -> unit Lwt.t) -> outcome:Types.test_outcome -> toolname:string -> 'raw_test corpus -> 'raw_test -> Stdlib.Digest.t Lwt.t

share_test ~on_share ~outcome ~toolname corpus v promises to share a given test with value v, generated by a tool named toolname, in corpus. If provided, on_share is called immediately with the identifier computed from v.

val share_test' : ?on_share:(Stdlib.Digest.t -> unit Lwt.t) -> outcome:Types.test_outcome -> toolname:string -> 'raw_test corpus -> 'raw_test -> unit Lwt.t

Version of share_test that discards the returned digest.

val on_new_test : 'raw_test corpus -> ('raw_test Types.test_view -> unit Lwt.t) -> (unit -> unit Lwt.t) Lwt.t

on_new_test corpus f promises to call f with the identifier and value of any new test that is given to corpus (e.g, via share_test). The promised/returned function must be called at most once to terminate the registration of f.

val stop_receiving_tests : _ corpus -> unit Lwt.t

stop_receiving_tests corpus returns a promise that is fulfilled whenever every test previously shared in corpus has been received and registered.

TODO: specify what happens if share_test is called afterwards.

val register_one_bypassed_test : _ corpus -> unit Lwt.t

register_bypassed_test corpus registers an additional test test whose coverage has been accounted for in the status store, but that is not inserted into the database.

val test_struct : _ corpus -> Sc_values.Struct.typ

val info : _ corpus -> Types.info

val has_crashes : Types.info -> bool

val has_oracle_failures : Types.info -> bool

val is_covering_outcome : Types.test_outcome -> bool

val is_covering_metadata : Types.test_metadata -> bool

val is_covering_test : _ Types.test_view -> bool

val is_triggering_rte_outcome : Types.test_outcome -> bool

val is_triggering_rte_metadata : Types.test_metadata -> bool

val is_triggering_rte_test : _ Types.test_view -> bool

val is_oracle_failure_outcome : Types.test_outcome -> bool

val is_oracle_failure_metadata : Types.test_metadata -> bool

val is_oracle_failure_test : _ Types.test_view -> bool

val compare_tests_by_serialnum : (_ Types.test_view as 'a) -> 'a -> int

module Types : sig ... end

module Printer : sig ... end

Pretty-printers for public types of this library.

module Decoder : sig ... end

module Validator : sig ... end

module Seeding : sig ... end

module Sharing : sig ... end