Module Sc_corpus.Sharing

val with_bidirectional_channel : ?import_suff:string -> ?import_filter:(Types.test_metadata -> bool Lwt.t) -> ?write_test: [< `Func of _ Sc_sys.File.TYPES.file -> 'raw_test -> unit Lwt.t | `Link Link ] -> share:('a Sc_sys.File.TYPES.file -> unit Lwt.t) -> 'raw_test Sc_corpus__.Main.corpus -> Sc_sys.File.TYPES.dir -> (unit -> 'a Lwt.t) -> 'a Lwt.t

with_bidirectional_channel ?import_suff ?import_filter ?write_test ~share corpus dirname f calls f after having setup a bi-directional exchange channel with corpus.

According to the underlying exchange protocol, each file newly created in directory dirname and whose name does not terminate with import_suff (".imported" by default), is registered and passed to share.

Reciprocally, the file of any test that: (i) is shared with corpus during the execution of f and until with_bidirectional_channel terminates, and (ii) whose metadata passes import_filter (if provided), is placed into dirname by using write_test: the latter is `Link by default, which means a filesystem link is used; alternatively a user-defined function write may be specified with `Func write. Files imported in this way are not subject to the aforementioned export mechanism.

val import_tests : ?import_suff:string -> ?write_test: [< `Func of _ Sc_sys.File.TYPES.file -> 'raw_test -> unit Lwt.t | `Link Link ] -> ?filter:(Types.test_metadata -> bool) -> 'raw_test Sc_corpus__.Main.corpus -> Sc_sys.File.TYPES.dir -> Basics.Digests.t Lwt.t

import_tests ?import_suff ?write_test ?filter corpus dirname imports every input previously (given to/shared with) corpus into the directory dirname. Inputs are imported locally using write if write_test = `Func write, or using a file-system link otherwise (i.e, write_test = `Link, the default).

filter may be used to select tests to import based on their metadata (cf Types.test_metatada).

The directory dirname must exist when this function is called.